+++ src/lib/components/Button/LoadingSpinner.svelte
@@ -1,12 +1,12 @@
 <script>
 	import Icon from '@iconify/svelte';
 
-	export let iconWidth = '1.3em';
+	export let iconWidth = '1em';
 	export let color = 'inherit';
 </script>
 
 <div class="rotating">
-	<Icon icon="ion:refresh-circle" width={iconWidth} {color} />
+	<Icon icon="tabler:loader-2" width={iconWidth} {color} />
 </div>
 
 <style>