+++ src/lib/components/Forms/InputWrapper.svelte
@@ -1,7 +1,7 @@
 <script type="ts">
 	import Icon from '@iconify/svelte';
 	import HelperMessage from './HelperMessage.svelte';
-	import { fly } from 'svelte/transition';
+	import { fade } from 'svelte/transition';
 	import TooltipIcon from '../TooltipIcon/TooltipIcon.svelte';
 
 	export let name: string;
@@ -45,15 +45,15 @@
 		</div>
 	{/if}
 	{#if errors.length > 0 && statusIcons}
-		<div class="icon-wrapper-right" transition:fly|local={{ x: 4, duration: 400 }}>
+		<div class="icon-wrapper-right" transition:fade|local={{ duration: 400 }}>
 			<Icon icon="tabler:alert-circle" color="var(--clr-alert-main)" width="0.9em" />
 		</div>
 	{:else if isValid && statusIcons}
-		<div class="icon-wrapper-right" transition:fly|local={{ x: 4, duration: 400 }}>
+		<div class="icon-wrapper-right" transition:fade|local={{ duration: 400 }}>
 			<Icon icon="tabler:check" color="var(--clr-success-main)" width="0.9em" />
 		</div>
 	{:else if pending && statusIcons}
-		<div class="icon-wrapper-right" transition:fly|local={{ x: 4, duration: 400 }}>
+		<div class="icon-wrapper-right" transition:fade|local={{ duration: 400 }}>
 			<Icon icon="tabler:loader-2" color="var(--clr-tertiary-main)" class="rotate" width="0.9em" />
 		</div>
 	{/if}