+++ src/lib/components/TransactionModal/TransactionModal.svelte
@@ -1,183 +1,231 @@
 <script type="ts">
-	import { fly } from 'svelte/transition';
+	import Icon from '@iconify/svelte';
+	import { createEventDispatcher } from 'svelte';
+	import { fly, fade } from 'svelte/transition';
 	import TransactionModalMessage from './TransactionModalMessage.svelte';
-	import { backInOut } from 'svelte/easing';
-	import { Column } from '@mateoroldos/svelte.bones';
 
-	let duration = 2000;
-	let easing = backInOut;
-	let options = { duration, easing, times: 1 };
-	function spin(node, options) {
-		const { times = 1 } = options;
-		return {
-			...options,
-			// The value of t passed to the css method
-			// varies between zero and one during an "in" transition
-			// and between one and zero during an "out" transition.
-			css(t) {
-				// Svelte takes care of applying the easing function.
-				const degrees = 100 * times; // through which to spin
-				return `transform: translate(${(t - 1) * -10}px) rotate(${
-					(t - 1) * degrees
-				}deg); opacity: ${1 * t};`;
-			}
-		};
-	}
-
-	interface FlowTransaction {
+	/** https://developers.flow.com/tools/fcl-js/reference/api#transactionstatusobject */
+	type TransactionStatusObject = {
 		blockId: string;
+		events: EventObject[];
+		status: TrnsactionStatus;
+		statusString: '' | 'UNKNOWN' | 'PENDING' | 'FINALIZED' | 'EXECUTED' | 'SEALED' | 'EXPIRED';
 		errorMessage: string;
-		events: {}[];
-		status: number | '';
-		statusCode: number | '';
-		statusString: string;
-	}
+		statusCode: string;
+	};
 
-	export let transactionLogo = '/flow-logo.png';
+	/** https://developers.flow.com/tools/fcl-js/reference/api#event-object */
+	type EventObject = {
+		blockId: string;
+		blockHeight: number;
+		blockTimestamp: string;
+		type: string;
+		transactionId: string;
+		transactionIndex: number;
+		eventIndex: number;
+		data: unknown;
+	};
+
+	type TrnsactionStatus = -1 | 0 | 1 | 2 | 3 | 4 | 5;
+
+	const dispatch = createEventDispatcher();
+
 	export let dappLogo = '/flow-logo.png';
 	export let transactionName = 'Flow';
-	export let transactionStatus: FlowTransaction;
+	export let transactionStatus: TransactionStatusObject;
 	export let transactionInProgress = false;
-	export let border = false;
+	export let border = true;
+
+	$: error = transactionStatus.errorMessage && Number(transactionStatus.statusCode) === 1;
+	$: success = transactionStatus.status === 4 && Number(transactionStatus.statusCode) === 0;
 </script>
 
 {#if transactionInProgress}
-	<article transition:fly={{ x: 100, duration: 800 }} class:border>
-		<Column gap="medium">
-			<Column gap="small">
-				<h6>{transactionName}</h6>
-				<span class="subtitle">Transaction</span>
-			</Column>
-			<div class="logos">
-				<div class="flow-logo pulse" transition:fly={{ y: -15, duration: 1800, delay: 200 }}>
-					<img src={transactionLogo} alt="Flow Logo" />
-				</div>
-				<div class="dapp-icon-container rotate" transition:spin={options}>
-					<img src={dappLogo} alt="Dapp Logo" />
+	<div class="main-wrapper">
+		<article transition:fly={{ x: 100, duration: 800 }} class:border class="column-10 align-center">
+			{#if error}
+				<button on:click={() => dispatch('close')} class="close">
+					<Icon icon="tabler:x" />
+				</button>
+			{/if}
+
+			<div class="column align-center">
+				<span class="subtitle small">{transactionName}</span>
+				<h4 class="h5">Transaction</h4>
+			</div>
+
+			<div class="logo-wrapper" class:error class:success>
+				<img class="pulse" src={dappLogo} alt="Dapp Logo" />
+				<div class="logo-border" />
+				<div
+					class="inverse-rotation-wrapper"
+					class:rotate-reverse={error}
+					class:rotate-slow={success}
+				>
+					<div class="logo-border secondary rotate" in:fade|local={{ duration: 500 }} />
 				</div>
 			</div>
-			<Column align="center" justify="center" gap="medium">
-				{#if transactionName === 'IPFS'}
+
+			{#if transactionName === 'IPFS'}
+				<TransactionModalMessage
+					title="Uploading Assets"
+					description="Uploading your assets to IPFS."
+				/>
+			{:else if transactionName === 'Flow'}
+				{#if transactionStatus.status < 0}
 					<TransactionModalMessage
-						title="Uploading Assets"
-						description="Uploading your assets to IPFS."
-						progressMessage="Uploading..."
+						title="Initializing"
+						description="Waiting for transaction approval."
 					/>
-				{:else if transactionName === 'Flow'}
-					{#if transactionStatus.status < 0}
-						<TransactionModalMessage
-							title="Initializing"
-							description="Waiting for transaction approval."
-							progressMessage="Initializing..."
-						/>
-					{:else if transactionStatus.status < 2}
-						<TransactionModalMessage
-							title="Pending"
-							description="The transaction has been received by a collector but not yet finalized in a block."
-							progressMessage="Executing"
-						/>
-					{:else if transactionStatus.status === 2}
-						<TransactionModalMessage
-							title="Finalized"
-							description="The consensus nodes have finalized the block that the transaction is included in."
-							progressMessage="Executing..."
-						/>
-					{:else if transactionStatus.status === 3 && transactionStatus.statusCode === 0}
-						<TransactionModalMessage
-							title="Executed"
-							description="The execution nodes have produced a result for the transaction."
-							progressMessage="Sealing..."
-							progress="80"
-						/>
-					{:else if transactionStatus.status === 4 && transactionStatus.statusCode === 0}
-						<TransactionModalMessage
-							title="Sealed"
-							description="The verification nodes have verified the transaction, and the seal is included in the latest block."
-							progressMessage="Sealing..."
-							progress="100"
-							icon="ion:checkmark-circle"
-						/>
-					{:else if transactionStatus.status === 5 && transactionStatus.statusCode === 0}
-						<TransactionModalMessage
-							title="Expired"
-							description="The transaction was submitted past its expiration block height."
-							progress={false}
-						/>
-					{:else if transactionStatus.errorMessage && transactionStatus.statusCode === 1}
-						<TransactionModalMessage
-							title="Failed"
-							description={transactionStatus.errorMessage}
-							progress={false}
-							icon="ion:close-circle"
-							error={true}
-						/>
-					{:else}
-						<TransactionModalMessage
-							title="Error"
-							description="An error occured."
-							progress={false}
-							error={true}
-						/>
-					{/if}
+				{:else if transactionStatus.status < 2}
+					<TransactionModalMessage
+						title="Pending"
+						description="The transaction has been received by a collector but not yet finalized in a block."
+					/>
+				{:else if transactionStatus.status === 2}
+					<TransactionModalMessage
+						title="Finalized"
+						description="The consensus nodes have finalized the block that the transaction is included in."
+					/>
+				{:else if transactionStatus.status === 3 && Number(transactionStatus.statusCode) === 0}
+					<TransactionModalMessage
+						title="Executed"
+						description="The execution nodes have produced a result for the transaction."
+					/>
+				{:else if transactionStatus.status === 4 && Number(transactionStatus.statusCode) === 0}
+					<TransactionModalMessage
+						title="Sealed"
+						description="The verification nodes have verified the transaction, and the seal is included in the latest block."
+						icon="tabler:circle-check-filled"
+					/>
+				{:else if transactionStatus.status === 5 && Number(transactionStatus.statusCode) === 0}
+					<TransactionModalMessage
+						title="Expired"
+						description="The transaction was submitted past its expiration block height."
+					/>
+				{:else if transactionStatus.errorMessage && Number(transactionStatus.statusCode) === 1}
+					<TransactionModalMessage
+						title="Failed"
+						description="Transaction execution failed."
+						error={true}
+					/>
+				{:else}
+					<TransactionModalMessage title="Error" description="An error occured." error={true} />
 				{/if}
-			</Column>
-		</Column>
-	</article>
+			{/if}
+		</article>
+	</div>
 {/if}
 
 <style type="scss">
-	article {
-		position: fixed;
-		top: 90px;
-		right: 20px;
-		background-color: var(--clr-primary-main);
-		box-shadow: 0 0 6px 0px var(--clr-primary-strong);
-		color: var(--clr-font-text-inverse);
-		padding: 2rem;
+	.main-wrapper {
 		z-index: 9999;
-		border-radius: 1.6rem;
-		width: 25ch;
+		position: fixed;
+		top: 0;
+		left: 0;
+		right: 0;
+		bottom: 0;
+		background-color: rgba(0, 0, 0, 0.654);
+		top: 0;
 
-		h6 {
-			font-size: var(--fs-400);
-			font-weight: 400;
-			text-transform: uppercase;
-			letter-spacing: 0.2ch;
-			color: var(--clr-primary-soft-t2);
-			margin: 0;
-			max-width: 20ch;
-			text-align: center;
-		}
-		.subtitle {
-			font-size: var(--fs-200);
-			font-weight: 300;
-			text-transform: uppercase;
-			letter-spacing: 0.5ch;
-			color: var(--clr-primary-soft-t6);
-			margin: 0;
-			max-width: 20ch;
-			text-align: center;
-		}
+		article {
+			position: fixed;
+			top: 110px;
+			right: 20px;
+			background-color: var(--clr-surface-secondary);
+			padding: 2rem;
+			z-index: 9999;
+			border-radius: var(--radius-7);
+			width: 25ch;
 
-		.logos {
-			position: relative;
+			.close {
+				position: absolute;
+				top: 6px;
+				right: 8px;
+				padding: 0.5rem;
+				background-color: transparent;
+				border: none;
+				cursor: pointer;
+				color: var(--clr-text-off);
+				transition: color 0.2s ease-in-out;
 
-			.flow-logo {
-				border-radius: 999rem;
-				width: 120px;
-				padding: 0;
+				&:hover {
+					color: var(--clr-heading-main);
+				}
 			}
 
-			.dapp-icon-container {
-				border-radius: 9999px;
-				position: absolute;
-				bottom: -12px;
-				right: -16px;
-				border: 4px var(--clr-primary-main) solid;
-				box-sizing: border-box;
-				box-shadow: 0 0 10px 1px var(--clr-primary-soft-t7);
-				width: 50px;
-				height: 50px;
+			.subtitle {
+				color: var(--clr-text-off);
+			}
+
+			.logo-wrapper {
+				position: relative;
+
+				.logo-border {
+					border-radius: 50%;
+					width: 120px;
+					height: 120px;
+					padding: 10px;
+					border: 7px solid var(--clr-tertiary-badge);
+
+					&.secondary {
+						position: absolute;
+						top: 0;
+						bottom: 0;
+						margin-left: auto;
+						margin-right: auto;
+						margin-top: auto;
+						margin-bottom: auto;
+						background-image: url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' rx='100' ry='100' stroke='#ff66c4' stroke-width='14' stroke-dasharray='20%25%2c 80%25' stroke-dashoffset='86' stroke-linecap='round'/%3e%3c/svg%3e");
+						border-radius: 100px;
+						border: none;
+						transition: all 0.5s ease-out;
+					}
+				}
+
+				img {
+					position: absolute;
+					left: 0;
+					right: 0;
+					top: 0;
+					bottom: 0;
+					margin-left: auto;
+					margin-right: auto;
+					margin-top: auto;
+					margin-bottom: auto;
+					border: 4px solid var(--clr-tertiary-badge);
+					border-radius: 50%;
+					width: 70px;
+					height: 70px;
+				}
+
+				&.error {
+					.logo-border {
+						border-color: var(--clr-alert-badge);
+
+						&.secondary {
+							background-image: url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' rx='100' ry='100' stroke='#FF5C5C' stroke-width='14' stroke-dasharray='20%25%2c 80%25' stroke-dashoffset='86' stroke-linecap='round'/%3e%3c/svg%3e");
+						}
+					}
+
+					img {
+						border-color: var(--clr-alert-badge);
+					}
+				}
+
+				&.success {
+					.logo-border {
+						border-color: var(--clr-primary-badge);
+
+						&.secondary {
+							background-image: url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' rx='100' ry='100' stroke='#3AE8C6' stroke-width='14' stroke-dasharray='20%25%2c 80%25' stroke-dashoffset='86' stroke-linecap='round'/%3e%3c/svg%3e");
+						}
+					}
+
+					img {
+						border-color: var(--clr-primary-badge);
+					}
+				}
 			}
 		}
 	}
@@ -200,30 +248,49 @@
 		animation-name: pulse;
 		animation-iteration-count: infinite;
 		animation-timing-function: ease-in-out;
-		animation-duration: 1.8s;
-		animation-delay: 1s;
+		animation-duration: 8s;
+		animation-delay: 2s;
 	}
 
 	@keyframes rotate {
-		33% {
-			-webkit-transform: rotate(20deg);
-		}
-		66% {
-			-webkit-transform: rotate(-20deg);
-		}
 		100% {
-			-webkit-transform: rotate(0deg);
+			-webkit-transform: rotate(360deg);
 		}
 	}
 
 	.rotate {
 		animation-name: rotate;
 		animation-iteration-count: infinite;
-		animation-timing-function: ease-out;
-		animation-duration: 3s;
+		animation-timing-function: linear;
+		animation-duration: 3.5s;
+		animation-direction: normal;
+	}
+
+	.inverse-rotation-wrapper {
+		width: 120px;
+		height: 120px;
+		position: absolute;
+		top: 0;
+		right: 0;
+
+		&.rotate-reverse {
+			animation-name: rotate;
+			animation-iteration-count: infinite;
+			animation-timing-function: linear;
+			animation-duration: 3.2s;
+			animation-direction: reverse;
+		}
+
+		&.rotate-slow {
+			animation-name: rotate;
+			animation-iteration-count: infinite;
+			animation-timing-function: linear;
+			animation-duration: 4.6s;
+			animation-direction: reverse;
+		}
 	}
 
 	.border {
-		border: 2px var(--clr-background-secondary) solid;
+		border: 1px var(--clr-border-primary) solid;
 	}
 </style>