+++ src/lib/models/navElement.interface.ts
@@ -3,4 +3,5 @@ export interface NavElement {
 	url: string;
 	prefetch: boolean;
 	icon?: string;
+	notifications?: number;
 }