+++ src/routes/+page.svelte
@@ -11,15 +11,18 @@
 	import FlowConnection from '$lib/components/FlowConnection/FlowConnection.svelte';
 	import Accordion from '$lib/components/Accordion/Accordion.svelte';
 	import CodeBlock from '$lib/components/CodeBlock/CodeBlock.svelte';
+	import ProgressSteps from '$lib/components/ProgressStep/ProgressSteps.svelte';
+	import Button from '$lib/components/Button/Button.svelte';
+	import TransactionModal from '$lib/components/TransactionModal/TransactionModal.svelte';
 
 	let stepss = [
 		{
 			name: 'First step',
-			state: 'active'
+			state: 'success'
 		},
 		{
 			name: 'Second step',
-			state: 'inactive'
+			state: 'loading'
 		},
 		{
 			name: 'Third step',
@@ -29,13 +32,26 @@
 
 	let value = 5;
 	let file: File[];
+
+	let dummyTransaction = {
+		blockId: '',
+		status: '4',
+		statusString: 'SEALED',
+		statusCode: '1',
+		errorMessage: '',
+		events: []
+	};
 </script>
 
 <section>
+	<button on:click={() => (dummyTransaction.errorMessage = 'aa')}>a</button>
+	<TransactionModal transactionInProgress={true} transactionStatus={dummyTransaction} />
 	<div class="container-small column-6">
+		<Button state="loading" statusIconsPosition="right">Loading</Button>
 		<a href="dwd" class="card-primary">
 			<h1>Card</h1>
 		</a>
+		<ProgressSteps steps={stepss} direction="column" />
 		<!-- <label for="a">
 			<input type="checkbox" id="a" />
 			Hola
@@ -189,7 +205,7 @@
 <section class="container-small">
 	<CodeBlock
 		codeBlockTitle="Cadence"
-		codeStyle="swift"
+		codeStyle="js"
 		code={`
 		pub contract Counter {
    			pub var count: Int